+++ docs/start/changelog.rst
@@ -11,6 +11,11 @@ ChangeLog
         $ lifx 'lan(default_broadcast="10.1.1.255"):get_attr' _ color
    * Added a default ``chain`` plan for the Gatherer to use for getting tile
      chain information
+   * FromGenerator can now be given a ``error_catcher_override`` option which
+     is a function that takes in the ``reference`` being operated on and the
+     original ``error_catcher``. It must return an ``error_catcher``. This can
+     be used with say ``FromGeneratorPerSerial`` to generate an ``error_catcher``
+     specifically for each serial.
 
 0.24.3 - 18 December 2019
     * Fixed a bug where response packets were matched to the wrong requests